+++ collector/src/test/java/com/bitmovin/analytics/ssai/SsaiEngagementMetricsServiceTest.kt
@@ -55,6 +55,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `markAdStart should createSampleWithStartedFlag`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
         ssaiEngagementMetricsService.flushCurrentAdSample()
 
@@ -71,12 +72,14 @@ class SsaiEngagementMetricsServiceTest {
         assertThat(adEventData.adId).isEqualTo("testId")
         assertThat(adEventData.adSystem).isEqualTo("testAdSystem")
         assertThat(adEventData.adIndex).isEqualTo(0)
+        assertThat(adEventData.adPodPosition).isEqualTo(0)
         assertThat(adEventData.adPosition).isEqualTo("midroll")
     }
 
     @Test
     fun `markAdStart should flush out existing sample before creating new one`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 1)
 
@@ -86,6 +89,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `markAdStart should be noop if adEngagementTracking is disabled`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsServiceDisabled.adBreakStart()
         ssaiEngagementMetricsServiceDisabled.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
         verify(exactly = 0) { analytics.sendAdEventData(any()) }
     }
@@ -93,6 +97,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `markAdStart should create new AdId on every call`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 1)
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 2)
@@ -225,6 +230,7 @@ class SsaiEngagementMetricsServiceTest {
 
     @Test
     fun `flushCurrentAdSample sends out existing sample and clears it afterwards`() {
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(null, null, 0)
 
         // first call should send out sample
@@ -244,6 +250,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `sendAdErrorSample should sendout existing sample with error data`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
         ssaiEngagementMetricsService.markQuartileFinished(SsaiAdPosition.MIDROLL, SsaiAdQuartile.FIRST, adMetadata, null, 0)
         ssaiEngagementMetricsService.sendAdErrorSample(
@@ -281,6 +288,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `sendAdErrorSample should NOT sendout sample if ssaiEngagement is disabled`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsServiceDisabled.adBreakStart()
         ssaiEngagementMetricsServiceDisabled.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
         ssaiEngagementMetricsServiceDisabled.sendAdErrorSample(
             SsaiAdPosition.MIDROLL,
@@ -297,6 +305,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `sendAdErrorSample is not send twice on one ad`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
         ssaiEngagementMetricsService.sendAdErrorSample(
             SsaiAdPosition.MIDROLL,
@@ -320,6 +329,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `sendAdErrorSample is called once per ad`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
         ssaiEngagementMetricsService.sendAdErrorSample(
             SsaiAdPosition.MIDROLL,
@@ -345,6 +355,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `timeout for flushing is reset on ad started`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
 
         verify(exactly = 2) { handlerMock.removeCallbacksAndMessages(any()) }
@@ -354,6 +365,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `timeout for flushing is postponed on quartile call`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
 
         // we clear the mock to only test that calls happen due to the quartile call, and not the started call
@@ -368,6 +380,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `timeout for flushing is cleared on error`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
 
         // we clear the mock to only test that calls happen due to the error, and not the started call
@@ -388,6 +401,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `timeout for flushing is cleared on completed`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
 
         // we clear the mock to only test that calls happen due to the completed call, and not the started call
@@ -400,6 +414,7 @@ class SsaiEngagementMetricsServiceTest {
     @Test
     fun `timeout for flushing is cleared on flushOfCurrentSamples`() {
         val adMetadata = SsaiAdMetadata(adId = "testId", adSystem = "testAdSystem")
+        ssaiEngagementMetricsService.adBreakStart()
         ssaiEngagementMetricsService.markAdStart(SsaiAdPosition.MIDROLL, adMetadata, 0)
 
         // we clear the mock to only test that calls happen due to the adBreakEnd call, and not the started call